requirements.txt
numpy>=1.21.0
argparse
pillow>=8.0.0
matplotlib>=3.4.0
pathlib>=1.0.1
torch>=1.13.0
torchvision>=0.14.0
scikit-image>=0.19.0
